requirements.txt
astroid==2.3.3
Django==2.2
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-js-asset==1.2.2
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
